修正number字段长度过短导致电话号码不能保存的问题

+++ contact.h
@@ -4,7 +4,7 @@ struct Person
 {
     char name[BUFFER_LEN];
     int age;
-    int number;
+    long long number;
     char email[BUFFER_LEN];
     char address[BUFFER_LEN];
 };
